Cleanup, but mainly we now use gluster.org prebuilt binaries
debian, ubuntu and centos docker images all run fairly old
versions of gluster FS by default.

So, we'll pull the latest build from gluster.org instead.

One tricky part was we need to include the packge apt-transport-https,
otherwise, we fail to do an https download via apt-get.

Ubuntu appears not to package apt-transport-https independently,
rather it uses a symlink to point to the http protocol in
/usr/lib/apt/methods.